test: more test adjustments
